authorization check on delete post

+++ controllers/feed.js
@@ -146,6 +146,11 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not Authorised!');
+                error.statusCode = 403;
+                throw error;
+            }
             clearImage(post.imageUrl);
             post.deleteOne();
         })
